Update
Navigation:

The navigation bar has been updated to refer users to the correct location
Navigation links have been updated as follows:

Home     Listings       Sell Your Home       New Home Buyer      Mortgage Calculator     Contact Us

The Home page has been updated to show a more clean look and feel.
The slider has been changed to a still image.

The search form has been removed and now shows two buttons leading to their respective pages.

Recent Listings have been added.
This is a container with 12 listings within a 3 x3 grid.

Looking To Sell A Home, Testimonials, and Partner Components have been added to the home page.

Buying a Home Page:

Buying a Home Page has been added and now gives users a plethora of information on the Services that Pacific North Offers.

The page also showcases a slider of featured homes that will potentially catch the users attention.

The Page includes a contact form as well as extra information on the Pacific North Team

Selling a Home PAge:

The Selling A Home Page has been added and includes the same amount of information and more as the Buying A Home Page.

This page also includes a contact form that allows the user to enter their address to be added to the list of data to be sent for Pacific North's evaluation.

Mortgage Calculator

A Mortgage Calculator has been added to the site

The Mortgage Calculator can estimate the Monthly Payment Breakdown for A home on the market.

With this calculator, you can enter your home price, down payment, interest rate and select a loan term that will be used to calculate your monthly mortgage and display it on a small graph.

Finally,

The Contact us Page

The Contact Us page has been updated to be slightly more user-friendly, giving users the ability to just click or tap on an icon of a way to contact Pacific North.

Testimonials have been included underneath the contact form to further ensure high-quality workmanship and cooperation.